refactor: Host Guest 컴포넌트 분리
